# Add specialized implementations of fixed stream message to reduce all…