--- a/core/src/main/java/com/linecorp/armeria/common/AggregatedHttpMessage.java
+++ b/core/src/main/java/com/linecorp/armeria/common/AggregatedHttpMessage.java
@@ -458,7 +458,7 @@ public interface AggregatedHttpMessage {
         if (!trailingHeaders.isEmpty()) {
             objs[writerIndex] = trailingHeaders;
         }
-        return FixedHttpRequest.ofHeadersWritten(headers, objs);
+        return FixedHttpRequest.ofWrittenHeaders(headers, objs);
     }
 
     /**