…ocations for common cases.

It's very common to allocate stream messages with 0-2 objects, so it's nice to have specializations to avoid the extra garbage of an object array. 

Unfortunately, the duplicate implementations introduce a lot of complexity. Let me know if you think it's worth it. `FixedHttpRequest` turned out much worse than I had originally expected.

The total allocation of each is
- Empty: 16 bytes (3 fields in base class)
- One element: 16 bytes (3 fields in base, one in sub)
- Two elements: 24 bytes (3 fields in base, 3 in sub)
- Three elements: 24 + 36 (3 fields in base, 3 in sub, 24 bytes array overhead, 12 bytes in array)

Also unrelated, removed `invokedOnSubscribe` as it's unnecessary for fixed streams which have no writes.

Benchmarks. Not totally apple-to-apple due to the `switch` overhead in `After` from what I can tell - I think the overhead is shown by `num=3`, where the implementation hasn't really changed, so presumably `After` should be treated a little higher than they look. But still pretty minor, let me know what you think.

```
After

Benchmark                            (flowControl)  (num)          (streamType)   Mode  Cnt         Score        Error  Units
StreamMessageBenchmark.jmhEventLoop          false      0  FIXED_STREAM_MESSAGE  thrpt   20  11644168.631 ± 152677.634  ops/s
StreamMessageBenchmark.jmhEventLoop          false      1  FIXED_STREAM_MESSAGE  thrpt   20   8573895.575 ±  34659.073  ops/s
StreamMessageBenchmark.jmhEventLoop          false      2  FIXED_STREAM_MESSAGE  thrpt   20   6602223.767 ±  24179.698  ops/s
StreamMessageBenchmark.jmhEventLoop          false      3  FIXED_STREAM_MESSAGE  thrpt   20   5326205.531 ±  17208.362  ops/s

Before

Benchmark                            (flowControl)  (num)          (streamType)   Mode  Cnt         Score       Error  Units
StreamMessageBenchmark.jmhEventLoop          false      0  FIXED_STREAM_MESSAGE  thrpt   20  11338855.960 ± 56332.436  ops/s
StreamMessageBenchmark.jmhEventLoop          false      1  FIXED_STREAM_MESSAGE  thrpt   20   8165172.831 ± 22656.208  ops/s
StreamMessageBenchmark.jmhEventLoop          false      2  FIXED_STREAM_MESSAGE  thrpt   20   6681994.146 ± 32692.861  ops/s
StreamMessageBenchmark.jmhEventLoop          false      3  FIXED_STREAM_MESSAGE  thrpt   20   5441368.221 ± 20874.088  ops/s

```