@anuraaga 2017-12-07
Thanks a lot for the suggestion! Indeed, there was significant duplication between those factories and the one in `HttpRequest`, and there's no good point for it. I made `FixedHttp*` package-private, and also fixed my misunderstanding that it's possible to have a `HttpRequest` that publishes headers (and made an explicit check to prevent this).

I also went ahead and deprecated `AggregatedHttpMessage.to*` in the interest of making `HttpRequest` and `HttpResponse` the only entrypoints for creating them (will make the API clearer after #850 ). In the process, I moved some error-checking logic from there to `HttpRequest/Response`, presumably it was supposed to be there either way.

Also, what do you think about moving `FixedStreamMessage` to `armeria.internal.stream`? I would only do it if we're also ok with moving `Default/EventLoopStreamMessage` there to later.