@trustin 2017-12-08
> Also, what do you think about moving `FixedStreamMessage` to `armeria.internal.stream`? I would only do it if we're also ok with moving `Default/EventLoopStreamMessage` to there later.

What's the benefit of moving them there? I usually prefer keeping classes package-local if possible since the classes in the internal package are visible to users anyway.